desing : 디테일 페이지 플레이리스트 부분 추가
